+++ inst/tinytest/test_duckdb_runtime_concurrent_queue.R
@@ -20,6 +20,7 @@ local({
 
   invisible(rducks_register(con, "rducks_queue_plus_one", function(x) x + 1, DOUBLE, DOUBLE))
   invisible(rducks_register(con, "rducks_queue_plus_one_rc", function(x) x + 1, DOUBLE, DOUBLE, eval_mode = "RC"))
+  invisible(rducks_register(con, "rducks_queue_plus_one_vec", function(x) x + 1, DOUBLE, DOUBLE, mode = "vectorized", side_effects = TRUE))
   rducks_enable_inproc(con)
 
   queued_result <- DBI::dbGetQuery(con, "SELECT sum(rducks_queue_plus_one(i::DOUBLE)) AS x FROM rducks_parallel_range(10::UBIGINT) AS t(i)")
@@ -27,6 +28,9 @@ local({
 
   queued_rc_result <- DBI::dbGetQuery(con, "SELECT sum(rducks_queue_plus_one_rc(i::DOUBLE)) AS x FROM rducks_parallel_range(10::UBIGINT) AS t(i)")
   expect_equal(queued_rc_result$x, sum((0:9) + 1))
+
+  queued_vec_result <- DBI::dbGetQuery(con, "SELECT sum(rducks_queue_plus_one_vec(i::DOUBLE)) AS x FROM rducks_parallel_range(10::UBIGINT) AS t(i)")
+  expect_equal(queued_vec_result$x, sum((0:9) + 1))
 
   final <- rducks_inproc_stats(con)
   expect_true(final$submitted[[1L]] > after$submitted[[1L]])